Move breadcrumbFromPathname function into the Breadcrumb component folder
